Add an alternative implmentation of the typesignature solver
An alternative implementation of the solver in dialyzer_typesig has
been introduced. It is faster than the original implementation.

Note: there is code for "loop detection". Where a loop occurs, the
evaluation is stopped and the current solution returned. This
behaviour is consistent with how the original implementation works.

There are a few known cases where the loop detection kicks in. They
are due to bugs which will hopefully be fixed in a near future.